MREADME.md
The objective of this project is to create an application “Theft Monitoring” where we are able to find our lost phone easily. 
Sometimes we accidently tend to replace our phone somewhere and we end up searching for it. Yet sometimes we are lucky to get our phones back but sometimes we never get it back.
Hence to avoid such situations and keeping the recent demand and the greed for phones in mind we have decided to develop this application “theft monitoring”.
This application gets activated as soon as the person who has stolen the phone changes the sim and switches on the phone. The user will not know that an automatic message is being sent to the predefined number.
This message will contain the sim details as well as the location details. This information will be sent as a message to the destined number till the application is uninstalled in the mobile. This will be like a torture to the user and makes it easy to find the location of the phone and track the user.
The location of the device is found using GPS and the details are sent to the predefined number in the form of a message as soon as the user changes the sim and switches on the phone.
The application also consists of the setting of the number to which the message regarding the sim and the location details are to be sent.
In this application all the operation is done in the background using the concept called “service”.
Thus “theft monitoring” is a useful application which helps us to track and find the device in an effective manner.